chore: worked on the gigify api

+++ src/api/v1/talents/services/talentPortfolio/deleteTalentPortfolio/index.ts
@@ -1,16 +1,26 @@
-import { BadRequestError, ControllerArgs, HttpStatus } from '@/core';
+import { BadRequestError, ControllerArgs, HttpStatus, ForbiddenError, UnAuthorizedError } from '@/core';
 import { TalentPortfolioParamDTO } from '~/talents/interfaces';
-import { TalentPortfolioRepository } from '~/talents/repository';
+import { TalentPortfolioRepository, TalentRepository } from '~/talents/repository';
 
 export class DeleteTalentPortfolio {
-    constructor(private readonly talentPortfolioRepository: TalentPortfolioRepository) {}
+    constructor(private readonly talentPortfolioRepository: TalentPortfolioRepository, private readonly talentRepository: TalentRepository) {}
 
     handle = async (payload: ControllerArgs<TalentPortfolioParamDTO>) => {
-        const { params } = payload;
+        const { params, request } = payload;
 
         if (!params) throw new BadRequestError(`Invalid Talent Portfolio ID`);
 
+        const userId = request.user?.id;
+
+        if (!userId) throw new UnAuthorizedError('User not authenticated');
+
         const { talentPortfolioId } = params;
+        const talent = await this.talentRepository.findByUserId(userId);
+        const portfolio = await this.talentPortfolioRepository.findById(talentPortfolioId);
+
+        if (!portfolio) throw new BadRequestError('Talent Portfolio not Found!');
+        if (!talent?.id) throw new UnAuthorizedError(`Talent not found! You're not registered as a talent.`);
+        if (portfolio.talent_id !== talent.id) throw new ForbiddenError('You do not have access to this portfolio item');
 
         await this.talentPortfolioRepository.deleteTalentPortfolio(talentPortfolioId);
 
@@ -21,6 +31,6 @@ export class DeleteTalentPortfolio {
     };
 }
 
-const deleteTalentPortfolio = new DeleteTalentPortfolio(new TalentPortfolioRepository());
+const deleteTalentPortfolio = new DeleteTalentPortfolio(new TalentPortfolioRepository(), new TalentRepository());
 
 export default deleteTalentPortfolio;
